Update MapKit to 4.42.0-lite and the whole toolchain (#46)
MapKit 4.41 stopped retaining listeners: on Android subscriptions take a
WeakReference, on iOS a __weak pointer. The wrapper mirrors that contract now.

- add WeakRef<T> (expect/actual) with asWeakRef() and withValue()
- add platform NativeConvertible<T> and WeakRef<T>.toNative()
- every subscription takes WeakRef<Listener>
- retain the cluster listener in ClusterNode, it used to be a factory local
- setMapFps takes Int, calculateCameraPosition takes focusRect before azimuth
  and tilt, IconStyle gained opacity, MapMode.HYBRID and
  LocationPurpose.STATIC_DISPLAY_LOCATION added

Toolchain: Gradle 9.6.1, AGP 9.3.1, Kotlin 2.4.10, Compose Multiplatform 1.11.1
and the rest of the catalog. AGP 9 forbids com.android.library next to the
Kotlin Multiplatform plugin, so the build was restructured:

- library modules use com.android.kotlin.multiplatform.library
- sample:composeApp is a KMP library, the Android app moved to sample:androidApp
- iosX64 dropped, Compose Multiplatform no longer publishes for it
- compileSdk/targetSdk 37, JVM target 11
- atomicfu Gradle plugin dropped, only its locks runtime API was used

MAPKIT_BACKLOG.md lists the 4.25-4.42 SDK API that is still not wrapped.

+++ yandex-mapkit-kmp-compose/src/commonMain/kotlin/ru/sulgik/mapkit/compose/MapConfig.kt
@@ -76,9 +76,9 @@ public data class MapConfig(
     /**
      * The base map type with [MapWindow.setMapFps].
      *
-     * Valid range: (0, 60]
+     * Valid range: [0, max display refresh rate]. 0 means max display refresh rate.
      *
      * If null use map's default value
      */
-    public val fps: Float? = null,
+    public val fps: Int? = null,
 )
